Use Handler.postDelayed instead of AlarmManager.setExact for Action.Auto
This is because as for Android >= 5.1 the minimum delay of AlarmManager.setExact()
changed to five seconds. so even when autoplay_span is less than five
seconds, it took five seconds to play next poem.

The historical reason why we used AlarmManager.setExact was to wake up
the device without adding WAKE_LOCK permission.

Therefore this commit will disable auto play when device goes to sleep.
However, I will try to wake up device in next commit.

But maybe it's a time to add WAKE_LOCK permission to this app instead of
using setExact to wake up device.
https://developer.android.com/training/scheduling/wakelock.html